+++ search-algorithms/binary-search/c/src/binsearch.h
@@ -1,6 +1,8 @@
 #ifndef BINSEARCH_H_
 #define BINSEARCH_H_
 
-int binsearch(const int haystack[], const int needle, const int h_size);
+#include <stddef.h>
+
+int binsearch(const int haystack[], const int needle, const size_t h_size);
 
 #endif // BINSEARCH_H_
